+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.6.0] - 2020-10-05
+
+- Added option to remove all "unused" component items through recursion
+- Improved the openapi-format terminal log output and how verbose is handled
+
 ## [1.5.0] - 2020-10-05
 
 - Added option to remove "unused" component items.